increase default mem when GPU task
