+++ sky/clouds/kubernetes.py
@@ -1,10 +1,8 @@
 """Kubernetes."""
 import json
-import math
 import os
-import re
 import typing
-from typing import Dict, Iterator, List, Optional, Tuple, Union
+from typing import Dict, Iterator, List, Optional, Tuple
 
 from sky import clouds
 from sky import exceptions
@@ -22,126 +20,6 @@ if typing.TYPE_CHECKING:
 logger = sky_logging.init_logger(__name__)
 
 _CREDENTIAL_PATH = '~/.kube/config'
-
-
-class KubernetesInstanceType:
-    """Class to represent the "Instance Type" in a Kubernetes.
-
-    Since Kubernetes does not have a notion of instances, we generate
-    virtual instance types that represent the resources requested by a
-    pod ("node").
-
-    This name captures the following resource requests:
-        - CPU
-        - Memory
-        - Accelerators
-
-    The name format is "{n}CPU--{k}GB" where n is the number of vCPUs and
-    k is the amount of memory in GB. Accelerators can be specified by
-    appending "--{a}{type}" where a is the number of accelerators and
-    type is the accelerator type.
-
-    CPU and memory can be specified as floats. Accelerator count must be int.
-
-    Examples:
-        - 4CPU--16GB
-        - 0.5CPU--1.5GB
-        - 4CPU--16GB--1V100
-    """
-
-    def __init__(self,
-                 cpus: float,
-                 memory: float,
-                 accelerator_count: Optional[int] = None,
-                 accelerator_type: Optional[str] = None):
-        self.cpus = cpus
-        self.memory = memory
-        self.accelerator_count = accelerator_count
-        self.accelerator_type = accelerator_type
-
-    @property
-    def name(self) -> str:
-        """Returns the name of the instance."""
-        assert self.cpus is not None
-        assert self.memory is not None
-        name = (f'{common_utils.format_float(self.cpus)}CPU--'
-                f'{common_utils.format_float(self.memory)}GB')
-        if self.accelerator_count:
-            name += f'--{self.accelerator_count}{self.accelerator_type}'
-        return name
-
-    @staticmethod
-    def is_valid_instance_type(name: str) -> bool:
-        """Returns whether the given name is a valid instance type."""
-        pattern = re.compile(r'^(\d+(\.\d+)?CPU--\d+(\.\d+)?GB)(--\d+\S+)?$')
-        return bool(pattern.match(name))
-
-    @classmethod
-    def _parse_instance_type(
-            cls,
-            name: str) -> Tuple[float, float, Optional[int], Optional[str]]:
-        """Parses and returns resources from the given InstanceType name
-
-        Returns:
-            cpus | float: Number of CPUs
-            memory | float: Amount of memory in GB
-            accelerator_count | float: Number of accelerators
-            accelerator_type | str: Type of accelerator
-        """
-        pattern = re.compile(
-            r'^(?P<cpus>\d+(\.\d+)?)CPU--(?P<memory>\d+(\.\d+)?)GB(?:--(?P<accelerator_count>\d+)(?P<accelerator_type>\S+))?$'  # pylint: disable=line-too-long
-        )
-        match = pattern.match(name)
-        if match:
-            cpus = float(match.group('cpus'))
-            memory = float(match.group('memory'))
-            accelerator_count = match.group('accelerator_count')
-            accelerator_type = match.group('accelerator_type')
-            if accelerator_count:
-                accelerator_count = int(accelerator_count)
-                accelerator_type = str(accelerator_type)
-            else:
-                accelerator_count = None
-                accelerator_type = None
-            return cpus, memory, accelerator_count, accelerator_type
-        else:
-            raise ValueError(f'Invalid instance name: {name}')
-
-    @classmethod
-    def from_instance_type(cls, name: str) -> 'KubernetesInstanceType':
-        """Returns an instance name object from the given name."""
-        if not cls.is_valid_instance_type(name):
-            raise ValueError(f'Invalid instance name: {name}')
-        cpus, memory, accelerator_count, accelerator_type = \
-            cls._parse_instance_type(name)
-        return cls(cpus=cpus,
-                   memory=memory,
-                   accelerator_count=accelerator_count,
-                   accelerator_type=accelerator_type)
-
-    @classmethod
-    def from_resources(cls,
-                       cpus: float,
-                       memory: float,
-                       accelerator_count: Union[float, int] = 0,
-                       accelerator_type: str = '') -> 'KubernetesInstanceType':
-        """Returns an instance name object from the given resources.
-
-        If accelerator_count is not an int, it will be rounded up since GPU
-        requests in Kubernetes must be int.
-        """
-        name = f'{cpus}CPU--{memory}GB'
-        # Round up accelerator_count if it is not an int.
-        accelerator_count = math.ceil(accelerator_count)
-        if accelerator_count > 0:
-            name += f'--{accelerator_count}{accelerator_type}'
-        return cls(cpus=cpus,
-                   memory=memory,
-                   accelerator_count=accelerator_count,
-                   accelerator_type=accelerator_type)
-
-    def __str__(self):
-        return self.name
 
 
 @clouds.CLOUD_REGISTRY.register
@@ -254,7 +132,7 @@ class Kubernetes(clouds.Cloud):
             cpus.strip('+')) if cpus is not None else cls._DEFAULT_NUM_VCPUS
         instance_mem = float(memory.strip('+')) if memory is not None else \
             instance_cpus * cls._DEFAULT_MEMORY_CPU_RATIO
-        virtual_instance_type = KubernetesInstanceType(instance_cpus,
+        virtual_instance_type = kubernetes_utils.KubernetesInstanceType(instance_cpus,
                                                        instance_mem).name
         return virtual_instance_type
 
@@ -263,7 +141,7 @@ class Kubernetes(clouds.Cloud):
         cls,
         instance_type: str,
     ) -> Optional[Dict[str, int]]:
-        inst = KubernetesInstanceType.from_instance_type(instance_type)
+        inst = kubernetes_utils.KubernetesInstanceType.from_instance_type(instance_type)
         return {
             inst.accelerator_type: inst.accelerator_count
         } if (inst.accelerator_count is not None and
@@ -273,7 +151,7 @@ class Kubernetes(clouds.Cloud):
     def get_vcpus_mem_from_instance_type(
             cls, instance_type: str) -> Tuple[Optional[float], Optional[float]]:
         """Returns the #vCPUs and memory that the instance type offers."""
-        k = KubernetesInstanceType.from_instance_type(instance_type)
+        k = kubernetes_utils.KubernetesInstanceType.from_instance_type(instance_type)
         return k.cpus, k.memory
 
     @classmethod
@@ -311,7 +189,7 @@ class Kubernetes(clouds.Cloud):
 
         # resources.memory and cpus are None if they are not explicitly set.
         # We fetch the default values for the instance type in that case.
-        k = KubernetesInstanceType.from_instance_type(resources.instance_type)
+        k = kubernetes_utils.KubernetesInstanceType.from_instance_type(resources.instance_type)
         cpus = k.cpus
         mem = k.memory
         # Optionally populate accelerator information.
@@ -381,7 +259,7 @@ class Kubernetes(clouds.Cloud):
             acc_type, acc_count = list(accelerators.items())[0]
 
             # Parse into KubernetesInstanceType
-            k8s_instance_type = KubernetesInstanceType.from_instance_type(
+            k8s_instance_type = kubernetes_utils.KubernetesInstanceType.from_instance_type(
                 default_instance_type)
 
             gpu_task_cpus = k8s_instance_type.cpus
@@ -389,7 +267,7 @@ class Kubernetes(clouds.Cloud):
             gpu_task_memory = float(resources.memory.strip('+')) \
                 if resources.memory is not None \
                 else gpu_task_cpus * self._DEFAULT_MEMORY_CPU_RATIO_WITH_GPU
-            chosen_instance_type = KubernetesInstanceType.from_resources(
+            chosen_instance_type = kubernetes_utils.KubernetesInstanceType.from_resources(
                 gpu_task_cpus, gpu_task_memory, acc_count, acc_type).name
 
         # Check if requested instance type will fit in the cluster.
@@ -421,7 +299,7 @@ class Kubernetes(clouds.Cloud):
         return {_CREDENTIAL_PATH: _CREDENTIAL_PATH}
 
     def instance_type_exists(self, instance_type: str) -> bool:
-        return KubernetesInstanceType.is_valid_instance_type(instance_type)
+        return kubernetes_utils.KubernetesInstanceType.is_valid_instance_type(instance_type)
 
     def validate_region_zone(self, region: Optional[str], zone: Optional[str]):
         # Kubernetes doesn't have regions or zones, so we don't need to validate
